requirements.txt
discord.py==1.6.0
python-dotenv==0.15.0
pytz==2019.3
PyNaCl==1.3.0
psycopg2-binary
pandas